+++ src/BuyListItem.jsx
@@ -3,7 +3,7 @@ const BuyListItem = ({product, deleteProduct, setProduct}) => {
 
     return(
       <div className="grid border border-gray-200 gap-x-4 rounded shadow p-4" style={{ gridTemplateColumns: "1fr 10fr 1fr 0fr 1fr 0fr 1fr 0fr"}}>
-        <img src="https://cdnimg.webstaurantstore.com/images/products/small/228802/1679695.jpg" alt="Refrigerator" className="p-2 rounded border border-gray-200" />
+        <img data-testid="product-image" src={product.image} alt="Refrigerator" className="p-2 rounded border border-gray-200" />
         <div className="flex flex-col flex-grow justify-center">
           <h3 data-testid="product-name" className="font-semibold text-lg">{product.name}</h3>
           <p>
@@ -23,7 +23,7 @@ const BuyListItem = ({product, deleteProduct, setProduct}) => {
 
         <span className="flex items-center">=</span>
 
-        <span className="flex items-center font-semibold text-green-500 text-xl">
+        <span data-testid="total-price" className="flex items-center font-semibold text-green-500 text-xl">
         ${(Math.round(product.price * product.quantity * 100) / 100).toFixed(2)}
         </span>
